[refactor] HttpStaticHandlerMapper 작성
HttpDynamicHandlerMapper 의 클래스 네이밍 변경에 따라 통일성을 위해서 같이 수정해줬습니다.
